pyproject.toml
[project]
name = "cv-course15"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.0",
    "matplotlib>=3.10.3",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "transformers>=4.53.3",
]
